Move DivvunSpellBannerCell to own file
Also rename it to DivvunSpellBannerCell
